+++ src/image_processing.py
